+++ test/integration/testing/functional.test.js
@@ -1,19 +1,34 @@
 /**
- * @overview Contains integration tests for the testing implementation of
+ * @overview Contains integration tests for the testing utilities provided with
  * Shescape.
  * @license MIT
  */
 
 import { testProp } from "@fast-check/ava";
+import test from "ava";
 import * as fc from "fast-check";
 
 import { arbitrary } from "../_.js";
 
 import { Shescape } from "shescape";
-import { Shescape as Stubscape } from "shescape/testing";
+import {
+  injectionStrings,
+  Shescape as Stubscape,
+  Throwscape,
+} from "shescape/testing";
+
+test("injection strings", (t) => {
+  t.true(Array.isArray(injectionStrings));
+  t.true(injectionStrings.length > 0);
+
+  for (const injectionString of injectionStrings) {
+    t.is(typeof injectionString, "string");
+    t.true(injectionString.length > 0);
+  }
+});
 
 testProp(
-  "escape (stubscape ~ shescape)",
+  "Stubscape#escape (stubscape =~ shescape)",
   [fc.anything(), arbitrary.shescapeOptions()],
   (t, arg, options) => {
     let result, stubResult, errored, stubErrored;
@@ -45,7 +60,7 @@ testProp(
 );
 
 testProp(
-  "escapeAll (stubscape ~ shescape)",
+  "Stubscape#escapeAll (stubscape =~ shescape)",
   [fc.anything(), arbitrary.shescapeOptions()],
   (t, args, options) => {
     let result, stubResult, errored, stubErrored;
@@ -77,11 +92,8 @@ testProp(
 );
 
 testProp(
-  "quote with shell (stubscape ~ shescape)",
-  [
-    fc.anything(),
-    arbitrary.shescapeOptions().filter((options) => options?.shell !== false),
-  ],
+  "Stubscape#quote, with shell (stubscape =~ shescape)",
+  [fc.anything(), arbitrary.shescapeOptions()],
   (t, arg, options) => {
     let result, stubResult, errored, stubErrored;
 
@@ -112,11 +124,8 @@ testProp(
 );
 
 testProp(
-  "quoteAll with shell (stubscape ~ shescape)",
-  [
-    fc.anything(),
-    arbitrary.shescapeOptions().filter((options) => options?.shell !== false),
-  ],
+  "Stubscape#quoteAll, with shell (stubscape =~ shescape)",
+  [fc.anything(), arbitrary.shescapeOptions()],
   (t, args, options) => {
     let result, stubResult, errored, stubErrored;
 
@@ -143,5 +152,16 @@ testProp(
 
     t.is(errored, stubErrored);
     t.is(typeof result, typeof stubResult);
+  },
+);
+
+testProp(
+  "Throwscape#constructor",
+  [arbitrary.shescapeOptions()],
+  (t, options) => {
+    t.throws(() => new Throwscape(options), {
+      instanceOf: Error,
+      message: "Can't be instantiated",
+    });
   },
 );